Introduce HistogramBase for common histogram behavior
